FCBHDBP-363 Added the way to synch the stage b and c parameters. It has added supporting to the method: getFilesetRecords10. It has merged the reader interface. Moved the LanguageRecord class to separated file. It has added supporting to execute the PreValidate using the stage C. It has update the way to fetch orthography value. It will point to the media.scriptCode column.
