+++ load/LanguageReaderCreator.py
@@ -21,10 +21,9 @@ class LanguageReaderCreator:
 
 if (__name__ == '__main__'):
     from Config import *    
-    languageReader = LanguageReaderCreator("B").create(Config().filename_lpts_xml)
-    #languageReader = LanguageReaderCreator("C").create()
+    config = Config()
+    migration_stage = os.getenv("DATA_MODEL_MIGRATION_STAGE") # Should be "B" or "C"
+    migration_stage = migration_stage if migration_stage != None else "B"
+    languageReader = LanguageReaderCreator(migration_stage).create(config.filename_lpts_xml)
     results = languageReader.getByStockNumber("N1SPA/PDT")
-    for duple in results:
-        print(duple)
-
-# python3 load/LanguageReaderCreator.py test
+    print("Reg_StockNumber [%s]" % results.Reg_StockNumber())
